requirements.txt
flask
google-generativeai
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dotenv
pytest
